refactor: move parse/compile style helpers to util file
