+++ rehype-plugins/gatsby-inline-images/index.js
@@ -2,22 +2,7 @@ const visit = require('unist-util-visit');
 const phrasing = require('hast-util-phrasing');
 const is = require('hast-util-is-element');
 const has = require('hast-util-has-property');
-
-const parseStyleString = (string) =>
-  string
-    .split(/;\s*/)
-    .filter(Boolean)
-    .reduce((style, rule) => {
-      const [property, value] = rule.split(/:\s*/);
-
-      return { ...style, [property]: value };
-    }, {});
-
-const compileStyleObject = (styles) =>
-  Object.entries(styles)
-    .filter(([, value]) => Boolean(value))
-    .map((rule) => rule.join(': '))
-    .join('; ');
+const { parseStyleString, compileStyleObject } = require('../utils/styles');
 
 const isGatsbyImageWrapper = (node) =>
   is(node, 'span') &&
